fix(input): horizontal overflow in IE and Edge (#2784)
* fix(input): horizontal overflow in IE and Edge

Currently input labels are set to 133% width and are scaled down to fit into their container. This works fine in most browsers, however it causes horizontal overflows on IE and Edge, because they don't consider the `transform` when determining whether an element overflows. This change fixes it by hiding all of the overflowing content.

* refactor: use alternative approach

* chore: re-add the fix after faulty merge