build: round payload size for comparison (#7174)
In some cases the payload is still showing up in the CI even if there was no change to the library. This seems to be because the file sizes are not rounded and can sometimes be not very exact. Resulting in numbers like: `material_fesm_2014": 1004.1319999999998`.

To have a better output in the GitHub statuses, the file size should be rounded