fix(drawer): infinite loop when two-way opened binding is toggled mid-animation (#8872)
As of https://github.com/angular/material2/commit/377602b3e4c4af353d37bd2dce30a41315b241d5 the drawer supports a two-way binding on the `opened` property, however because the `openedChange` emits at the end of the animation and two-way bindings invoke their setter which then trigger an animation which in turn invokes `openedChange`, there is the potential for an infinite loop if the user changes the `opened` state mid-animation. These changes avoid the issue by turning the `openedChanged` into an async stream and emitting the current value at the end of the animation, rather than determining the value based on the animation state.

Fixes #8869.