fix(overlay): attempting to position overlay if it was detached immediately after being attached (#9507)
Adds a check that ensures that the `OverlayRef` doesn't attempt to position itself if it was detached before the zone managed to stabilize. This manifested itself as an issue in the select where it could end up throwing an error, because the relevant nodes are no longer in the DOM.

Fixes #9406.