fix(menu): lazy-rendered content being duplicated when toggling quickly (#11348)
Fixes the lazily-rendered content of a menu being duplicated, if it user toggles the panel too quickly. The issue comes from the fact that we wait for the animation to finish before we detach the content, however we allow the panel to be reopened as soon the closing sequence is kicked off.

Fixes #11331.