perf(overlay): only compute and emit position changes if there are subscribers (#11431)
When emitting the `ConnectedOverlayPositionChange`, we have to compute whether the overlay is being clipped by the viewport or one of its parent scrollables. These calculations can be expensive, because they need the `ClientRect` of each element, however they won't necessarily be used by the consumer. These changes move things around so we only calculate the scrollable clipping if the consumer has subscribed to the `positionChanges` observable.