fix(chips): chip list removing focus from first chip when adding through the input (#12840)
Fix regression in 963c626d55ca7e7e58533e9d3cd509ec987356e0. Currently when the user removes all the chips and then they add a new chip, the chip list will remove focus from the input and put it on the chip. These changes introduce the proper behavior, which is to keep focus on the input.