fix(button): ripple color not correct for standard, icon and stroked buttons (#13235)
With https://github.com/angular/material2/commit/613f5e63098d708a8d82623a5428e4d94765f8ef, we already ensured that the text color is **inherited** in order to make sure that buttons without a background color are readable on various themed backgrounds (e.g. themed toolbars)

This is a follow-up that ensures that the ripple color for such transparent buttons is contrasting to the background color (e.g. in themed toolbars).

Fixes #13232