build: example list outputting quoted components (#13955)
In dde0043993f4b5b864ffd072ba94482c32fe5e78 the logic that outputs the list of components was reworked to use `JSON.stringify`. As a result, the data that is generated has the `component` inside quotes, rather than a reference to the component class itself. This breaks the demo app. The following changes add some extra logic to remove the quotes around the `component`.