fix(tabs): paginated header offset incorrect on IE (#13223)
Currently we position the paginated tab header items using a `transform` in order to be able to animate them. On IE setting the transform also ends up changing the scroll position of the parent element, causing the tab offset to be thrown off. These changes reset the scroll after setting the transform.

Fixes #13217.