fix(slide-toggle): thumb animation not working on mobile (#15236)
With https://github.com/angular/material2/commit/22b7e0509a0ee563ad68339e86b0605a69306159, we enabled the `grabbing` cursor if the thumb container is pressed. This accidentally broke animation for the thumb sometimes because the `:active`
pseudo class is added on tap/touch and now accidentally sets the thumb transition duration to `0ms`. Rather we should only set the thumb transition duration to `0ms` if the user started dragging the thumb (with a small threshold)

Fixes #15232