build: do not ship data typescript sources to npm (#15597)
In the early stages of the schematics the upgrade data was
stored in `.json` files. We switched these data files to TypeScript
at some point, but didn't update the schematic-copy glob. Meaning
that we currently ship TypeScript source files within the `data/` folder
to NPM for the CDK and Material schematics.