fix(table): text-column throws if name input is set eagerly with ivy
Currently when someone uses the `CdkTextColumn` directive and
sets the `name` input binding to a static value, Ivy will set the input
value eagerly in the creation phase where the static query results
are not available yet. This causes a runtime exception since the
`this.columnDef` variable is still set to undefined when the `name`
input setter executes.