chore: fix mdc-checkbox not working with ivy
Fixes that the `mdc-checkbox` prototype does not work properly with
Ivy as there are component inputs which are commonly set programmatically
but do not mark the checkbox as dirty.

This surfaced in one of the checkbox tests because with Ivy, calling `markForCheck`
from within the `ngAfterViewInit` lifecycle hook has basically no effect while View
Engine keeps the component dirty.

This caused a checkbox test failure because the test just assigned the
value of the `MatCheckbox#checked` property and called `detectChanges`
afterwards. Since the component is not marked as dirty due to the aforementioned
Ivy difference, the component will not be checked and updated. In order to fix
the test we just mark the component for check in the input setters (similarly to
how we do it for other inputs of `MatCheckbox` such as `disabled`).