fix(tabs): only use aria-current on active links (#16562)
We previously set `aria-current` to either "true" or "false" for each
`matTabLink`. This change now only applies the attribute to the active
tab, omitting it for inactive tabs. Additionally, instead of setting
true, we now set the attribute to "page", which specifically indicates
that the item is a link that refers to the current page.

Fixes #16557