refactor(checkbox): remove unused host binding for required validator (#16548)
Currently the `mat-checkbox` required validator defines a host binding
that adds the "required" attribute to the `mat-checkbox` host element.

This is not necessary as the `required` attribute should be only added
to the underlying input element (which is handled as part of the checkbox
template).

This unnecessary host binding seems to have been added by
accident since the `CheckboxRequiredValidator` from `@angular/forms`
has the same host binding. Through the difference is that the
`CheckboxRequiredValidator` from `@angular/forms` is guaranteed
to have native input elements as host element.