feat(material-experimental): add test harness for radio-group
Follow-up for 8f516990c41c11b12d5d27bef1df1c1da6a7f4f1. This commit
introduces a new test harness for the `mat-radio-group` implementation.

Note that we can't provide harness methods for getting the selected
value because the selected value or the value of a radio-button are
part of the component-internal state (no way to determine from DOM)