fix(radio-button): underlying input not respecting value binding
Currently a `mat-radio-button` does not respect the value binding that has
been set on the radio-button.

This is not ideal because in most cases the value is a primitive that can
be set as `value` for the underlying input element in order to ensure that
proper `FormData` can be retrieved when the form gets submitted.

Currently the form data will always have `on` as value because the selected
radio-button does not pass-through the value to the underlying input.

(cherry picked from commit bd3bf09abcb2ff2588dab8e56b3f50cf38e3f48c)