test(sidenav): refactor flaky tests in Edge (#16625)
Refactors some tests to (hopefully) be less flaky in Edge. In two of them I've replaced calls to `getComputedStyle` to a check against a class in the DOM. On other one I've changed the assertions for IE and Edge to something that is less accurate, but also less likely to flake.

(cherry picked from commit 7123e1f53d942dce5e9a8597a91465a8277e88f5)