feat(material-experimental): add menu test harness skeleton (#16615)
Adds a skeleton for the menu test harness. It currently captures the
basic functionality for the trigger. The functionality for the menu
panel/items will come in a follow-up PR. As part of this, I'm adding a
css class to the menu trigger so that it's findable via DOM.

(cherry picked from commit 2c97504766ec54c63f75f5024a29d1dabaeac7dc)