fix(drag-drop): unable to drop into connected sibling that was scrolled into view (#16681)
In the scenario where two drop lists are connected and one of them is scrolled out of view, even if the user scrolls it into view, the list won't accept the dragged item until the user starts a new drag sequence. The issue comes from the fact that we only update the cached dimensions of the list in which we're currently dragging, but not the ones that it's connected to. These changes fix the issue by also updating the dimensions of all connected lists.