build: mdc cronjob tests should use "develop" branch (#16991)
MDC seems to do their primary work in the "develop" branch. We
discussed in the team meeting that we want to run tests against
that branch as "master" is just their release staging branch.

The goal of the snapshot job is to detect issues as soon as
possible so that we can notify MDC in case of possible changes
that negatively impact us.