refactor(core): allow ElementRef in RippleRenderer.setupTriggerEvents (#16953)
The way the `RippleRenderer` is set up at the moment is inconsistent from an API standpoint, because the constructor only accepts an `ElementRef`, whereas `setupTriggerEvents` only accepts an `HTMLElement`, and in most cases they're referring to the same DOM node. These changes make the two code paths consistent with each other by allowing `HTMLElement | ElementRef<HTMLElement>` in both of them.