fix(list): selection list checkbox theme overwritten by parent theme (#16939)
Fixes the theme of the pseudo checkbox inside the `mat-list-option` being overwritten if the list is placed inside of an element that has a different default theme. The issue comes from the default checkbox styles not being specific enough and the selection list not setting a class for `mat-accent`.

Fixes #16891.

(cherry picked from commit 2cefb6977a1eaa81b6cdb7e8597d8a8eb2204428)