fix(a11y): focus monitor not identifying touch focus inside shadow root (#17167)
`FocusMonitor` uses a `touchstart` listener on the `document` to identify when an element was focused via touch. The problem is that if an element is inside the shadow DOM, the `event.target` will be set to the shadow root. These changes use `composedPath` to get the event target which accounts for the shadow DOM.

(cherry picked from commit 3673f3d3084b91e6f619b835a3431532c8e1909d)