build: fix gulp release output not building (#17248)
Even though we do not use the Gulp release output anymore, we should fix
it in case we need to roll-back the Bazel release output migration.

To fix the build, the flat module bundle for the Material primary entry-point has
been removed. We cannot use that anymore since there are no-exports in the
primary entry-point and having `index.ts` as entry-point and flat-module entry-point
the same time causes a NGC metadata bundle stack overflow exception (expected)

Additionally we no longer build the examples with Gulp since that migration was
already successful and the Gulp setup does not work with tertiary entry-points
which are extensively used in the examples package.