build: publish snapshot scripts incorrectly updating unrelated code (#17290)
I've been wondering for quite some time why a comment in the
cdk drag-drop changes sometimes in the `cdk-builds`.

Apparently it happens because the publish snapshot script used
the current version in the `package.json` as search pattern for
a sed call. This is breaky since the version includes a dot to
separate version segments. These dots are not interpreted as
dots, but rather as a RegExp special character.

Causing unrelated changes to comments that match e.g. /8.2.0/

https://github.com/angular/cdk-builds/commit/b3a86279410f3df409e1d1e4638b3b3ed20b8eef#diff-24a5156f00a6496d7fd08504682e8246R1952.