build: generate api docs for all tertiary entry-points (#17293)
Updates our Dgeni setup to generate API docs for all
entry-points which are configured in the `cdk/config.bzl`
or `material/config.bzl` file.

This basically means that we now generate API docs for
`material/*/testing` entry-points and for `cdk/testing/*`.

Removes the code that made `cdk/testing` a separate sub
package. This involves more code than just treating it
as "real" entry-point (for which we generate bundles etc.).