build: changelog should not show changes to google-maps (#17313)
We don't release the google-maps package and disabled the
package group in the changelog. Apparently there is a bug where
we do not detect commits as google-maps changes if the commit
format does not specify an explicit package, but rather just a scope.

We usually determine the package based on the scope by looking
for matching packages. This did not work for `google-maps` changes
such as 461ee8e