fix(drag-drop): not picking up indirect descendant items (#17226)
Fixes that the drop list doesn't pick up drop items that aren't direct descendants. This was put in on purpose before to try and enforce a simpler DOM structure, but it could be limiting in some cases (e.g. if users want to put an `ngSwitch` with an `ng-container` around it).

These changes also add some extra logic to ensure that items aren't picked up by more than one ancestor list.

Fixes #17047.