build: avoid bazel analysis cache discarding (#17461)
It looks like Bazel sometimes discards the analysis cache
when switching between testing/running and plain building.

This is because 91de590ea1f6eeffe85e53d7374caa570c6a6887 introduced a flag
for disabling legacy bazel behavior. This flag has been only
applied to "run" and "test", but not for "build". This means
that we accidentally switch the environment, and Bazel discards
the analysis cache (slowing down the development).