build: reduce manual setup for entry-points of packages (#17460)
Reduces the manual setup for entry-poitns of packages. With
this change, we no longer need to update the dev-app system
config if we add/remove entry-points.

We basically only need to configure entry-points in the
`config.bzl` file. Though this is not 100% complete yet,
because we still have legacy karma tests which need another
system config. The bazel karma tests do not need this anymore.