docs: add package ID prefix to h3's ids (#17527)
This helps avoid collisions in the table of content's sections like
"Directives" and "Services", since now we may show more than one
package entry point in docs (e.g. button and button-testing).

Note that we chose not to add the prefix to individual
class/constant/interface/etc names because it muddles the URL and
is unlikely that we collide on these values.