feat: expand input types with coercion to work with ngtsc input type checking (#17528)
* feat: expand input types with coercion to work with ngtsc input type checking

* refactor: add workaround for strict metadata emit of coercion static members

See: https://github.com/angular/angular/issues/33451

* build: patch ng_module bazel rule to enable strict template type checking

Workaround for: https://github.com/angular/angular/issues/33452

* refactor: disable strict metadata emit for legacy build output

Workaround for https://github.com/angular/angular/issues/33451. This makes
sense at the current time since the Bazel output does *never* work with
strict metadata emit... and considering that with Ivy there is no metadata
anyway, it could be disabled in the meanwhile.

* build: disable strict attribute type checking

We are temporarily disabling strict attribute type checking
since we have a few templates that set a coerced input while
the `NgModel` directive is applied. In those cases, the empty string
for the `disabled` input, is not assignable to the `NgModel#disabled` input,
since it does not do coercion. We need to figure out what to do in those
scenarios.