fix(bottom-sheet): don't move focus if it was moved during close animation (#17725)
Currently we restore focus to the previously-focused element once the bottom sheet's exit animation has completed, however this can override any focus management the consumer might have done while the animation was running. These changes add extra check so that we only move focus if it was inside the bottom sheet at the end of the animation.

Along the same lines as #17300.