fix(menu): cap maximum elevation for nested menus (#17687)
Currently we increase the elevation of each nested menu by one, however we don't have an upper limit which means that menus 20 levels down won't have any elevation (the elevation starts at 4). This is a bit of an edge case since opening 20 menus is a lot, but it's easy to handle on our end.