fix(drag-drop): position reset if viewport is resized while boundary is invisible (#17777)
If a dragged item is inside a boundary, we update its position on window resize to ensure that it stays inside the boundary. If a resize event happens while the boundary is hidden, we end up resetting the user's position unintentionally. These changes add a check that prevents the position from changing if the elements are hidden.

Fixes #17750.