fix(material-experimental/mdc-slide-toggle): fix extra border… (#18035)
In the latest MDC Canary version an extra border was added to the slide toggle. The border also gets set to `transparent`, but due to insertion order gets overwritten by the base shorthand `border` declaration. This was only an issue for the `primary` color, because we were assuming that it's the default so we included less specific styles. These changes make it the same specificity as the other colors.

Also bumps us to the latest MDC Canary version.