build: add tsickle as explicit dependency (#18116)
Adds tsickle as explicit dependency. This is necessary
because we access it from `node_modules/tsickle` in a
Bazel target that wraps tsc with tsickle.

In the past we just relied on the transitive dependency
of `@angular/bazel` being hoisted, but the hoisting is
not guaranteed if the lock file changes (e.g. when running `yarn
upgrade`). In general, dependencies we explicitly access, should be
added as explicit dependencies.