build: update rules_sass version to latest (#18242)
Updates `rules_sass` to the latest version that contains:

https://github.com/bazelbuild/rules_sass/commit/dd0079be310f918d50fe6a196ff82e5872433eb4.

This fixes that errors are reported silently and only shown
once the CSS file is served in the browser. It still means that
the message might not be printed in the terminal, since there
is still https://github.com/bazelbuild/rules_sass/issues/96, but
it's better seeing that something failed at compilation time.