fix(expansion): unable to toggle disabled panel via methods (#18181)
In our expansion panel readme we state that a disabled panel can still be toggled and this is partially true, because it works only for toggling through the `expanded` property, but not via the `open`, `close` and `toggle` methods since they're inherited from the CDK `AccordionItem`. These changes override the methods so that we can allow them to work on disabled panels. We can't make this change safely for the CDK accordion item, because it doesn't have a panel header which means that we have to assume that all calls to the methods are user-generated.

Fixes #18171.

(cherry picked from commit d350318d7f329917ef147bfa1910ddd7528674ff)