fix(tabs): better high contrast indication on supported browsers (#18160)
We add an outline for focus indication in high contrast mode for tabs, but part of it gets clipped, because of `overflow: hidden`. These changes add a negative offset to ensure that it doesn't clip. Note that this isn't supported on IE, but it's a progressive enhncement for Edge and Firefox.

(cherry picked from commit 745d519cffd35506419ca18e78e6ee75411bbfdd)