fix(tree): not allowing horizontal overflow (#18215)
The way the tree styles were set up initially means that the tree element can't cause its parent to overflow. Given that the tree can have a lot of nesting, this is bound to happen with deeper trees. These changes remove the default `overflow: hidden` since it doesn't seem to be doing much.

Fixes #18182.