fix(drag-drop): error if next sibling is removed after drag sequence has started (#18230)
When a drag sequence is started, we save a reference to the next sibling of the drag element so that we can use it at the end of the sequence to restore the element's original DOM position. This breaks down and throws an error if the next sibling gets removed while the user is dragging. These changes make the approach a bit more robust by inserting a comment node instead.

Fixes #18205.

(cherry picked from commit 0763fb9e0f6056321c1178f7a72a685b86f56dcf)