fix(checkbox): outline not visible when checked in high contrast mode (#18048)
Fixes checked checkboxes blending in with the background in high contrast mode. This is something that used to work, but might've regressed after we made `cdk-high-contrast` to generate a class selector which has a different specificity.

Also moves the checkbox high contrast styles into the main component styles so they don't have to be repeated for each theme.