fix(ng-add): do not overwrite version range specified in `ng add` (#18365)
* fix(ng-add): do not overwrite version range specified in `ng add`

The CLI inserts requested dependencies into the `package.json` before
`ng add` schematics run. This means that we usually do not need to insert
the dependency into `package.json` files.

Alternatively, it could happen that the `ng add` schematics run outside of
the CLI `ng add` command, or the requested dependency is dev only. In those
cases we insert a version based on the current version placeholder.

* fixup! fix(ng-add): do not overwrite version range specified in `ng add`

Address feedback