build: remove unused angular bazel postinstall patch (#18373)
In the past, when the framework repo updated their `rules_nodejs`
version, a patch on the components was needed in order to make the
`components-repo-unit-tests` job on framework green.

Since this update PR already landed, and we updated to the Angular
version that includes these changes, we can remove the patch.