feat(material-experimental/mdc-form-field): support for disabl… (#18397)
By leveraging the animations feature target, we can add support for
disabling animations of the form-field. Unlike in non experimental
Material components, the class for toggling the animations is now
scoped to the component. This gives more fine-grained control and
makes it public. The class doesn't need to be denoted as private, as
it can be useful for implementers of custom form-field controls.

Also fixes that the line-ripple styles are not included.