fix(material-experimental/mdc-form-filed): missing focus and hover effect for fill
We recently updated to the latest canary version of MDC, but did not
account for a new breaking change to the MDC text-field that required
us to explicitly specify a `ripple` container element.

https://github.com/material-components/material-components-web/commit/c541ebe157a66e8d2e881fad16cc4dbe30b2c16b#diff-fa38508bb43a471ee25b746d85527fb5

Thinking more about the ripple styles in the MDC text-field unveiled
that we should just provide our own implementation of the focus overlay
since MDC ripples are too much bloat for the simple case of showing
a focus and hover effect. We don't want to create a `MdcRipple` runtime
instance, nor do we want to include CSS selectors for interaction
ripples which are not used in the text-field at all.