refactor(material-experimental/mdc-form-field): use mdc-line-ripple for inactive bottom-line
Previously, MDC only set the bottom-line for text-fields on the input
and textarea elements. This didn't work for us since we support custom
form field controls. We added a custom bottom-line for this.

MDC recently made a change to move the bottom-line inactive styles to
the actual `mdc-bottom-line` (which we already use for active styles).

This means that we can get rid of our custom styles. Reference commit:
https://github.com/material-components/material-components-web/commit/b6c7f624bc7d88e2e371efcb125c7a6bac55eab7.