fix(ng-update): better detection for workspace project in v9 hammer migration (#18525)
Currently the v9 HammerJS migration determines the workspace project by
consulting the `ts.Program` that has been constructed. This logic is not
guaranteed to work because a TypeScript program doesn't necessarily need
to contain any "root" file names. This means that we cannot reliably
determine the workspace project from the `ts.Program` and we will throw
an error that no project could be found.

We can improve this logic by simply using the workspace project that is
associated with the originating tsconfig file. This was previously not
possible, but cb577b1c7c323695e7560d5e8f7bd1fed8be2c47 enabled us to
pass through the workspace project.

Fixes #18504.