refactor(dialog): remove custom injector usage (#18565)
Replaces the usage of our custom `PortalInjector` with Angular's `Injector`. The advantage of using the default one is that the injection logic is consistent with everything else and we don't have to require users to polyfill `WeakMap`.

The idea here is to eventually replace all usages of `PortalInjector` with `Injector` and to deprecate it, but I wanted to try it out against `dialog` since it's probably the most used and well tested of the components that are using the custom injector.

(cherry picked from commit ff8d56e8f137478486996541a09b1ef8ae7eb78b)