fix(chips): only add type attribute to button remove icons (#18477)
In #18095 we made it so that the chip remove button automatically adds `type="button"` so that it doesn't accidentally submit forms. It was made so that it always gets added since it was less code than adding extra checks, but the problem is that it ends up conflicting with some CSS resets. These changes make it so that we only add it for buttons.

Fixes #18464.

(cherry picked from commit 9badc54ccc1c5b1d8c42b86f21ead425cd6a68b0)