build: fix dev-server running in snippet mode (#18679)
Browsersync has 3 possible modes. `snippet`, `proxy` and `server`. We
currently run with `snippet` even though we actually want to serve with
the live-reloading script being injected automatically.

In order to ensure that Browsersync runs in the `server` mode and
properly prints the URL to the http server, we re-enable the
`server` option that has been disabled in the past for security reasons.
The logic to prevent directory traversal still remains, and we also
explicitly disable directory listing (even though we have manual request
interception).