fix(material-experimental/mdc-form-field): avoid expression ch… (#18741)
There is currently a bug in the Angular Framework that results in
ExpressionChangedAfterItHasBeenChecked errors not throwing when they
should for OnPush components. Becuase we modify a variable used in the
template after it was already checked (in ngAfterViewInit), we need to
trigger change detection manually or the change might not be picked up.