refactor: ensure typography mixins expect consistent arguments (#18738)
Currently we have a few components with typography mixins that do
not match the majority of other components. These components have
typography mixins where the `$config` is optional. We want to
always require a config (similar to how it's done in other components)
to avoid breaking changes in the future, and to have a consistent
API.