fix(radio): clicks not landing correctly in some cases on Chrome (#18357)
This is along the same lines as #18285 since the radio button and slide toggle have a very similar setup. Whenever the radio button host receives focus we forward it immediately to the underlying input, but this bouncing around of focus could cause clicks to be interrupted in some cases. These changes make it so that we only move focus if the host was focused by the keyboard or mouse.

(cherry picked from commit fe298350cd8391ce9a24b56c275c297bdf2b8a12)