fix(clipboard): not restoring focus to SVG elements (#18626)
When the clipboard is copying something, focus is temporarily moved to a `textarea` and moved back. The logic that moves it checks whether the previous element was an `HTMLElement`, but that doesn't cover returning focus to an SVG element.

(cherry picked from commit ddc9fd6baa4f75a8b2edc20b7bec9edf92353203)