fix(material/mdc-form-field): text-field should stretch based on host element (#18778)
Currently the MDC text-field uses `inline-flex` but does not expand
to the width of the form-field host element. We want the text-field
to be based on the actual form-field host element as it makes it easier
for developers to control the width of form-fields. i.e. adding
`display: block`, or specifying a maximum width.

(cherry picked from commit 26ab43bde3e002f0f3a5803e7c0a4b6d873ac85b)