feat: report warning if duplicate theme styles are generated
As of the new density API, a theme by default includes density styles
for the zero density-scale. This means that the common pattern of using
the `angular-material-theme` pattern for secondary themes in a project
causes duplicated styles for density.

We want to detect this case and report a warning. Most of the time,
for these secondary themes, only the color styles are desired. If this
is not the case and a special case surfaces here, the warning can be
ignored, but it's still helpful reporting that same styles are generated
multiple times.