refactor: add support density scaling in the expansion panel
Note: This commit is intentionally not marked as a feature because
density styles for the expansion panel are not public API yet:

Adds support for density scaling to the Angular Material
expansion panel. To be able to add support for density, the
panel header animation had to be reworked to be CSS based.

The existing inputs for setting collapsed/expanded height
still work as expected, but will not respect density.

* chore: fix buildifier formatting due to update

fix: unable to set expanded or collapsed height in expansion panel (#19322)

For the density system, we reworked the expansion panel to no
longer rely on Angular animations. Looks like the refactoring
missed that units are already set for the `expandedHeight` and
`collapsedHeight` inputs.

This hasn't been catched in any unit test, so tests are updated to catch
this now.