fix(drag-drop): preview matchSize sometimes incorrect inside flex container (#19062)
We create a custom preview through `ViewContainerRef.createEmbeddedView` and we move it to the correct place in the DOM. The problem is that for the brief period that it's in the DOM, we also measure the host node if the `matchSize` option is set. In some cases (e.g. a flex container), this can throw off the size. These changes switch to measuring the size ahead of time when necessary.

Fixes #19060.

(cherry picked from commit 5ef4fa5f4d77af858a97fd6850be26e7d0c5eeab)