fix(material-experimental/mdc-slide-toggle): strong focus indication not visible (#19048)
In #18895 a `border-radius: 50%` was added to the ripple host element of the MDC-based slide toggle in order to support different densities. The problem is that the ripple element is also used to draw a rectangle in strong focus indication mode. These changes move the strong focus indication to the `.mdc-switch__thumb-underlay` and switch the base focus indication to use `::after` instead of `::before`.

(cherry picked from commit 35efeabaa9ae9419364dd14393b0dbfb774a0952)