fix(select): don't select active item when tabbing away while closed (#18797)
When the user tabs away from a select, we pick out the active option automatically. The problem is that we weren't checking that the select is open so in some cases the user might be tabbing through and selecting something by accident.

Fixes #18784.

(cherry picked from commit 630183b1154be6c27fd63044cd009b9a699ef081)