fix(google-maps): incorrect variable access for server-side rendering check (#19166)
We added a SSR check to the google-maps window info directive, but when
this PR landed in the patch branch, the variable names no longer
matched. The variable for the `_infoWindow` has been made public
with 5f3303c80fd5e56b589b214801eef7779ac452db, but this change did not
land in the patch branch as it targets a next minor release.