build: update to typescript 3.9 (#19286)
The Framework team is in the process of updating to TS 3.9 (see https://github.com/angular/angular/pull/36989) which means that soon we'll have to switch to it as well. These changes update us to the latest RC and resolve all of the build issues so that we can quickly switch to the final version once it's available.