feat: do not generate default density styles multiple times
For backwards compatibility, we do not want to generate density
styles multiple times. With the new API, the theme mixins include
density styles by default and this could mean that common theme setup as
followed could result in duplicate density styles:

```
@include angular-material-theme($light-theme);

.dark-theme {
  @include angular-material-theme($dark-theme);
}
```

Ideally, developers would now use the `angular-material-color` mixin for
their dark theme styles (as recommended in the theming guide), but we
cannot expect this to happen for all applications immediately, so we
provide code in Angular Material that automatically ensures that the
default density styles are not generated a second time.

The goal is to generate default density styles only *once* and at
top-level if the legacy theming API is used. If density styles would
be duplicated through explicit use of the new theming API, we'll show a
warning and refer to the dedicated "duplicate styles" guide.