build: update to latest ts-node version (#19399)
The dev-infra package has an optional peer dependency
on `ts-node` for supporting TypeScript-based configurations.

This highlighted that we currently are behind the latest version
by a good amount. We should update so that the peer dependency
is satisified and generally for the sake of benefiting from the
latest `ts-node` changes (like better diagnostic messaging).

This commit additionally fixes compilation issues in the
caretaker-resync script and in the breaking changes script.

(cherry picked from commit 23beea39d49e541b8efe7c8e320492758f06b6b1)