fix(youtube-player): binding some event listeners too late (#19429)
The logic that lazily binds event listeners to the YouTube player was waiting for the player to be done initializing which is problematic if we want to bind the `onReady` listener since the event has already happened by the time we've bound it. These changes move the event binding logic earlier in the process.

Fixes #19408.