fix: focus monitor-based styles not working in some cases inside shadow dom (#19422)
A while ago event delegation was implemented for the focus monitor which has some special behavior where it has to bind its events to the closest shadow root in order for them to work correctly. The problem is that a lot of our `FocusMonitor.focus` calls are very early in the component lifecycle which means that might not have been put into the shadow DOM yet. These changes move the `monitor` calls to `ngAfterViewInit` to resolve the issue.

Fixes #19414.