build: setup slack notifications for circleci jobs (#19608)
Whenever a CI job fails in a non-PR build, the failure
will be reported to the `#components-ci-failures` Slack
channel.

Obviously this can become very verbose as it will report
every individual CI job, but that cannot be improved at the
time of writing, as CircleCI does not support post-workflow logic.

Interestingly there is also a legacy option in the old CircleCI UI
for Slack notifications, but for the sake of not relying on that,
we use the new recommended way of controlling notifications inside
the actual CircleCI configuration.