build: option to automatically retry postinstall patches (#19615)
Sometimes we need to change existing postinstall patches. This
requires a cleanup of the currently installed node modules, but
there is no good warning/messaging. We improve this with this
commit by detecting such stale patches and prompting for
automatic retry with cleaned up node modulesbuild: option to
automatically retry postinstall patches

Sometimes we need to change existing postinstall patches. This
requires a cleanup of the currently installed node modules, but
there is no good warning/messaging. We improve this with this
commit by detecting such stale patches and prompting for automatic
retry with cleaned up node modules.

Note: Please clean up your node modules after this change. This needs to
still happen manually once, so that we can keep track of applied
patch versions and report/prompt automatically in the future.