refactor: use lightweight token for CDK accordion
The `CdkAccordionItem` currently always brings in the `CdkAccordion`
class. This is not desired as an accordion item could be used
outside of a `<cdk-accordion>`. In those cases, we do not want
to retain the accordion class w/ metadata unnecessarily.

This also affects the `MatExpansionPanel` as it extends
from the `CdkAccordionItem` and therefore always brings in
the `CdkAccordion` list unnecessarily right now.

Related to #19576.