build: increase snapshot publish job queue confidence threshold (#19843)
We recently added a new step to the snapshot publish job. The
step will block the job until all previous snapshot publish
jobs that have been launched before complete.

This step relies on the CircleCI API which does not always
refresh immediatiely. This means that we could run into
a race condition where the queue does not detect previous
jobs. This has been observed with:

https://app.circleci.com/pipelines/github/angular/components/9618/workflows/52fef2b9-8059-4c00-8c2e-347b83ad1031/jobs/158528/steps

We fix this by increasing the confidence threshold of the
queue step, so that it will perform a second CircleCI request
after a delay to avoid the race condition in the most common
cases. It's a trade-off between stability and delay. We could
certainly increase the threshold further, but that will slow
down CI on publish branches, so we start experimenting with lower
threshold values first.