build: ensure project version is set correctly in NPM package output (#19841)
In framework, the version placeholder computation was incorrect
in some situations. i.e. the release packages used a version that
indicates local changes like `x.x.x-local-changes`. We should ensure
that the release output matches the expected version.

This is especially becoming relevant when we switch to a shared
Bazel workspace status command from the dev-infra package. The
workspace status command is responsible for determining the version
placeholder value when building with `--config=release`.