fix(cdk/a11y): interactivity checker detecting tabbable state incorrectly for videos, audio and object elements
The interactivity checker has been built a while ago when supported
browsers behaved differently for their `tabbable` state. We now
updated our browsers to more recent versions, and need to adjust
the interactivity checker so that it reflects the latest patterns
in supported browsers.

Here is a note summary of behavior noticed for `video` and `audio`
elements in various supported browsers:
https://hackmd.io/@devversion/rkxzcgsJD

Another change was for `object` elements. In previous BLINK / Webkit
browsers, such elements, or its children were never tabbable. This
is incorrect for the currently supported browsers as children inside
`object` elements can certainly be tabbable. On the other hand though,
object elements itself can be tabbable depending on configuration
we cannot reliably detect. For improved accessibility, we never
consider object elements itself as tabbable (similar to iframes)