fix(drag-drop): not transferring input value when cloning element for preview (#20009)
We use `cloneNode` to create a preview for an element which clones all of the attributes of an element, but that may not pick up properties like its `value`. These changes add some logic where we clone the value with a similar approach that we use for cloning `canvas` elements. Also generalizes the approach a bit so we don't have to keep repeating the logic.

Fixes #19905.

(cherry picked from commit fadbb3dddf076ed4ca83e253389d6b8e51138fb6)