fix(datepicker): not updating calendar state if input changes while open (#19973)
The datepicker was only updating the calendar state that depends on the input when we open and close. I've reworked it so we do it any time one of the input's inputs changes. I also renamed the various `_disabledChange` streams to `_stateChanges` which aligns with what we're doing in other components.

Fixes #19959.

(cherry picked from commit 0c1fccaa944a46b5dded3e4ad039b8047ca7115b)