fix(menu): throw error if menu contains its own trigger (#19943)
If the consumer puts a trigger inside a menu and passes the menu back to the trigger, we'll go into an infinite loop and throw a cryptic error message. These changes detect such a case and throw a better error. Example of a case that will be flagged:

```
<mat-menu #menu="matMenu">
  <button [matMenuTriggerFor]="menu"></button>
</mat-menu>
```

Fixes #19941.

(cherry picked from commit 4b6316ef994197c88f4efded5aaf90a88512e28e)