fix(select): indicating programmatic value change as user interaction in some cases (#19970)
We were using `ListKeyManager.setActiveItem` to change the active item when a value is assigned through a `FormControl`, in order to have it be highlighted if the user opens the panel. The problem is that `setActiveItem` emits a change event as if it's coming from the user, causing us to dispatch an event object with `isUserInput: true`.

These changes use `updateActiveItem` instead which only updates the item, but doesn't emit any events.

Fixes #19967.

(cherry picked from commit ec17c684b46606bd4e2c5d65ede74b2d8d3f5930)