fix(clipboard): scroll position changing while copying on some browsers (#20073)
The clipboard module works by creating a `textarea` dynamically, focusing it and copying its content. In order to prevent it from affecting the page layout, we put it at `-999em` off-screen, but the problem is that some browsers will try to move it into the view on focus by scrolling up.

These changes use `position: fixed` and anchor the textarea to the top of the page so vertically it's within the viewport.

(cherry picked from commit 9c538436ab42298d2387dc1ee77418fec38c8c7d)