fix(scrolling): virtual scroll not accounting for margin when measuring range (#19852)
Currently the `CdkVirtualForOf` determines the size of a rendered range by adding up the heights of all the elements, however this doesn't account for margins between them. These changes switch to doing it by taking the difference between the bottom of the last element and the top of the first. This should be a minor performance improvement as well, because we don't have to measure as many elements anymore.

Fixes #19851.

(cherry picked from commit e11e57f9d09b8e8cad48d3a5428ae80bc98a90c8)