test: clean up remaining dynamic query usages (#20156)
For a while we had linting to explicitly specify the `static` flag on queries and we eventually removed it in f15a907d16235bfa1773ba5432d5a943a4aaa872, but there are still some PRs going in that were written while the restriction was in place. These changes clean up the remaining instances.