build: create script for building docs-content (#20198)
Creates a script that builds the docs-content. We also use
that script in the our CI setup to reduce duplication.

The script will not pack the package by default, as that is
up to the consumer of the script. It's totally fine just
adding a package directory to a project (similar to a tarball).

(cherry picked from commit 53c9441b8ff11174d575fa84056c526caba72ff8)