fix(cdk/overlay): ensure OverlayOutsideClickDispatcher listens to contextmenu events (#20373)
If an overlay is open and a user opens a contextmenu (on an element outside the overlay) the
overlay does not close. This fix ensure that contextmenu events are treated the same way as click
events which result in the outsidePointerEvents observable emitting an event.

(cherry picked from commit 7a63ad69c147ad799ad0412e8d900213216b35f4)